finished but it still need a deeper look into it for validity and speed (more info in the description)
- finished the implementation and it is pretty slow in spitting out the
result, i suspect that is because of the horrendous stack implementation
i did 3 years ago, and the exesive copying of data at each stage of the
simulation tho i havent thought of a solution for that.
- about validity tho i am not sure, because in theory the more room the
simulation have the better it need to get and surpace the old algorithm,
but what happens is the opposite sometimes and incosistant, possibly
because of an error in the implementation or and error in the
theoritical approach of the algorithm, gonna investigate that more to
determine the real reason.